llm-bot-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
neo4j==5.14.1
langchain==0.0.350
openai==1.3.0 